custom hooks useDelete and useGetNodeData added
